feature: 주문저장 병합 및 사용자 주문조회 구현 (테스트코드 작성 필요)

+++ src/main/java/com/smplatform/product_service/domain/order/entity/OrderProduct.java
@@ -37,7 +37,7 @@ public class OrderProduct {
     @Column(name = "product_name")
     private String productName;
     @Column(name = "discount_id")
-    private Long discountId;
+    private Integer discountId;
     @Column(name = "discount_type")
     private Discount.Type discountType;
     @Column(name = "disocunt_value")
